WIP assets heist
