+++ front/app/routes/_main/map/$placeId_/heist/$heistId/delete/route.tsx
