+++ front/app/routes/_main/map/$placeId_/heist/$heistId/join/route.tsx
