+++ front/app/routes/_main/map/$placeId_/heist/$heistId/leave/route.tsx
