+++ front/app/routes/_main/map/$placeId_/heist/$heistId/prepare/PaymentDisplay.tsx
