+++ front/app/routes/_main/map/$placeId_/heist/add/route.tsx
@@ -1,20 +1,21 @@
 import * as Dialog from '@radix-ui/react-dialog';
-import { Button, Grid, Heading, Section } from '@radix-ui/themes';
+import { ArrowLeftIcon } from '@radix-ui/react-icons';
+import { Button, Flex, Grid, Heading, Section } from '@radix-ui/themes';
 import { json, redirect } from '@remix-run/node';
 import { useLoaderData } from '@remix-run/react';
 import { ClientError } from 'graphql-request';
+import { useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { RemixFormProvider, getValidatedFormData, useRemixForm } from 'remix-hook-form';
 
-import { getAssets } from '~/lib/api/asset';
+import { getGlobalAssets } from '~/lib/api/asset';
 import { getEmployeesEstablishments } from '~/lib/api/employee';
 import { getEstablishmentsOfContractor } from '~/lib/api/establishment';
-import { getHeist, heistIsMadeBy, heistIsPublic, updateHeist } from '~/lib/api/heist';
+import { createHeist } from '~/lib/api/heist';
 import { HeistDifficultyEnum, HeistPreferedTacticEnum, HeistVisibilityEnum } from '~/lib/api/types';
 import { getUsers } from '~/lib/api/user';
 import { Link } from '~/lib/components/Link';
 import { FormAlertDialog } from '~/lib/components/dialog/FormAlertDialog';
-import { SubmitButton } from '~/lib/components/form/SubmitButton';
 import { FieldInput } from '~/lib/components/form/custom/FieldInput';
 import { FieldInputArray } from '~/lib/components/form/custom/FieldInputArray';
 import { FieldMultiSelect } from '~/lib/components/form/custom/FieldMultiSelect';
@@ -22,102 +23,51 @@ import { FieldSelect } from '~/lib/components/form/custom/FieldSelect';
 import { TextAreaInput } from '~/lib/components/form/custom/TextAreaInput';
 import { i18next } from '~/lib/i18n/index.server';
 import { commitSession, getSession } from '~/lib/session.server';
-import { getMessageForErrorStatusCodes, hasErrorStatusCodes, hasPathError } from '~/lib/utils/api';
+import { getMessageForErrorStatusCodes, hasErrorStatusCodes } from '~/lib/utils/api';
 import dayjs from '~/lib/utils/dayjs';
 import { ROLES } from '~/lib/utils/roles';
 import { formatEnums } from '~/lib/utils/tools';
-import { updateHeistResolver } from '~/lib/validators/update-heist';
+import { createHeistResolver } from '~/lib/validators/create-heist';
 import { FLASH_MESSAGE_KEY } from '~/root';
-import { denyAccessUnlessGranted, hasRoles } from '~utils/security.server';
+import { denyAccessUnlessGranted } from '~utils/security.server';
 
 import type { ActionFunctionArgs, LoaderFunctionArgs } from '@remix-run/node';
 import type { Option } from '~/lib/types/select';
-import type { UpdateHeistFormData } from '~/lib/validators/update-heist';
+import type { CreateHeistFormData } from '~/lib/validators/create-heist';
 import type { FlashMessage } from '~/root';
 
-export async function loader({ context, params, request }: LoaderFunctionArgs) {
+export async function loader({ context, params }: LoaderFunctionArgs) {
   const user = denyAccessUnlessGranted(context.user, ROLES.CONTRACTOR);
-  const isAdmin = hasRoles(context.user, ROLES.ADMIN);
 
-  if (!params.heistId) {
-    throw redirect(`/map/${params.placeId}`);
-  }
+  // Get the establishments of the current user
+  const { establishments } = await getEstablishmentsOfContractor(context.client, user.id);
+  const establishmentsIds = establishments.edges.map((edge) => edge.node.id);
+  const [{ employees }, { assets }, { users }] = await Promise.all([
+    getEmployeesEstablishments(context.client, establishmentsIds),
+    getGlobalAssets(context.client),
+    getUsers(context.client),
+  ]);
 
-  const t = await i18next.getFixedT(request, 'response');
-
-  try {
-    const isPublic = await heistIsPublic(context.client, params.heistId);
-
-    if (isPublic) {
-      return redirect(`/map/${params.placeId}`);
-    }
-
-    const { heist } = await getHeist(context.client, params.heistId);
-
-    if (heist.establishment.contractor.id !== user.id && !isAdmin) {
-      return redirect(`/map/${params.placeId}`);
-    }
-
-    // Get the establishments of the current user
-    const { establishments } = await getEstablishmentsOfContractor(context.client, user.id);
-    const establishmentsIds = establishments.edges.map((edge) => edge.node.id);
-    const [{ employees }, { assets }, { users }] = await Promise.all([
-      getEmployeesEstablishments(context.client, establishmentsIds),
-      getAssets(context.client),
-      getUsers(context.client),
-    ]);
-
-    return {
-      user,
-      users,
-      heist,
-      assets,
-      employees,
-      establishments,
-      placeId: params.placeId,
-      locale: context.locale,
-    };
-  } catch (e) {
-    if (!(e instanceof ClientError) || !hasPathError(e, 'heist')) {
-      throw e;
-    }
-
-    throw new Response(null, {
-      status: 404,
-      statusText: t('not_found', { ns: 'response' }),
-    });
-  }
+  return {
+    user,
+    users,
+    assets,
+    employees,
+    establishments,
+    placeId: params.placeId,
+    locale: context.locale,
+  };
 }
 
 export type Loader = typeof loader;
 
 export async function action({ request, context, params }: ActionFunctionArgs) {
-  const user = denyAccessUnlessGranted(context.user, ROLES.CONTRACTOR);
-  const isAdmin = hasRoles(context.user, ROLES.ADMIN);
-
-  if (!params?.heistId) {
-    throw redirect(`/map/${params.placeId}`);
-  }
-
-  const isPublic = await heistIsPublic(context.client, params.heistId);
-
-  if (isPublic) {
-    return redirect(`/map/${params.placeId}`);
-  }
-
-  const isMadeBy = await heistIsMadeBy(context.client, {
-    id: params.heistId,
-    userId: user.id,
-  });
-
-  if (!isMadeBy && !isAdmin) {
-    throw redirect(`/map/${params.placeId}`);
-  }
+  denyAccessUnlessGranted(context.user, ROLES.CONTRACTOR);
 
   const t = await i18next.getFixedT(request, ['validators', 'flash']);
-  const { errors, data } = await getValidatedFormData<UpdateHeistFormData>(
+  const { errors, data } = await getValidatedFormData<CreateHeistFormData>(
     request,
-    updateHeistResolver,
+    createHeistResolver,
   );
 
   if (errors) {
@@ -141,20 +91,22 @@ export async function action({ request, context, params }: ActionFunctionArgs) {
 
   try {
     const { startAtTime, startAtDate, shouldEndAtDate, shouldEndAtTime, ...heistData } = data;
-    await updateHeist(context.client, {
+    await createHeist(context.client, {
       ...heistData,
-      id: params?.heistId,
       minimumPayout: +heistData.minimumPayout,
       maximumPayout: +heistData.maximumPayout,
       startAt: dayjs(`${startAtDate} ${startAtTime}`).toISOString(),
-      shouldEndAt: dayjs(`${shouldEndAtDate} ${shouldEndAtTime}`).utc(false).toISOString(),
+      shouldEndAt: dayjs(`${shouldEndAtDate} ${shouldEndAtTime}`).toISOString(),
+      visibility: HeistVisibilityEnum.Draft,
       allowedEmployees: heistData.allowedEmployees.map((allowedEmployee) => allowedEmployee.value),
-      forbiddenUsers: heistData.forbiddenUsers?.map((user) => user.value),
       forbiddenAssets: heistData.forbiddenAssets?.map((asset) => asset.value),
+      forbiddenUsers: heistData.forbiddenUsers?.map((user) => user.value),
+      objectives: heistData.objectives ?? [],
+      placeId: params.placeId,
     });
 
     session.flash(FLASH_MESSAGE_KEY, {
-      content: t('heist.updated_successfully', { ns: 'flash' }),
+      content: t('heist.created_successfully', { ns: 'flash' }),
       type: 'success',
     } as FlashMessage);
 
@@ -184,11 +136,10 @@ export async function action({ request, context, params }: ActionFunctionArgs) {
   );
 }
 
-export type Action = typeof action;
-
-export default function Edit() {
+export default function Add() {
   const { t } = useTranslation();
-  const { placeId, heist, employees, assets, users, user } = useLoaderData<Loader>();
+  const { placeId, establishments, employees, assets, users, user } = useLoaderData<Loader>();
+
   const usersFormatted = users.edges.reduce<Option[]>((acc, curr) => {
     if (user.id !== curr.node.id) {
       acc.push({
@@ -199,86 +150,99 @@ export default function Edit() {
 
     return acc;
   }, []);
+
   const assetsFormatted: Option[] = assets.edges.map((edge) => ({
     label: edge.node.name,
     value: edge.node.id,
   }));
-  const employeesFormatted = employees.edges.reduce<Option[]>((acc, curr) => {
-    if (heist.establishment.id === curr.node.establishment.id) {
-      acc.push({
-        label: curr.node.user.username,
-        value: curr.node.id,
-      });
-    }
 
-    return acc;
-  }, []);
-  const heistVisibilities = formatEnums(Object.values(HeistVisibilityEnum), 'heist.visibility');
-  const heistPreferedTactics = formatEnums(
-    Object.values(HeistPreferedTacticEnum),
-    'heist.prefered_tactic',
+  const employeesFormatted: (Option & { establishmentId: string })[] = employees.edges.map(
+    (edge) => ({
+      establishmentId: edge.node.establishment.id,
+      label: edge.node.user.username,
+      value: edge.node.id,
+    }),
   );
-  const heistDifficulties = formatEnums(Object.values(HeistDifficultyEnum), 'heist.difficulty');
-  const methods = useRemixForm<UpdateHeistFormData>({
+
+  const establishmentsFormatted: Option[] = establishments.edges.map((edge) => ({
+    label: edge.node.name,
+    value: edge.node.id,
+  }));
+
+  const heistPreferedTactics = formatEnums(Object.values(HeistPreferedTacticEnum));
+  const heistDifficulties = formatEnums(Object.values(HeistDifficultyEnum));
+
+  const dateNow = dayjs();
+  const startAt = dateNow.add(15, 'minutes');
+  const shouldEndAt = startAt.add(15, 'minutes');
+
+  const methods = useRemixForm<CreateHeistFormData>({
     mode: 'onSubmit',
-    resolver: updateHeistResolver,
-    submitConfig: {
-      unstable_viewTransition: true,
-    },
+    resolver: createHeistResolver,
     defaultValues: {
-      name: heist.name,
-      description: heist.description,
-      startAtDate: dayjs(heist.startAt).utc(false).format('YYYY-MM-DD'),
-      startAtTime: dayjs(heist.startAt).utc(false).format('HH:mm'),
-      shouldEndAtDate: dayjs(heist.shouldEndAt).utc(false).format('YYYY-MM-DD'),
-      shouldEndAtTime: dayjs(heist.shouldEndAt).utc(false).format('HH:mm'),
-      preferedTactic: heist.preferedTactic,
-      difficulty: heist.difficulty,
-      minimumPayout: heist.minimumPayout,
-      maximumPayout: heist.maximumPayout,
-      allowedEmployees: heist.allowedEmployees.edges.map((edge) => ({
-        value: edge.node.id,
-        label: edge.node.user.username,
-      })),
-      forbiddenUsers: heist.forbiddenUsers.edges.map((edge) => ({
-        value: edge.node.id,
-        label: edge.node.username,
-      })),
-      forbiddenAssets: heist.forbiddenAssets.edges.map((edge) => ({
-        value: edge.node.id,
-        label: edge.node.name,
-      })),
-      visibility: heist.visibility,
-      objectives: heist.objectives.map(
-        (objective: { name: string; description: string; optional?: boolean }) => ({
-          ...objective,
-          optional: objective.optional ?? false,
-        }),
+      startAtDate: startAt.format('YYYY-MM-DD'),
+      startAtTime: startAt.format('HH:mm'),
+      shouldEndAtDate: shouldEndAt.format('YYYY-MM-DD'),
+      shouldEndAtTime: shouldEndAt.format('HH:mm'),
+      establishment: establishments.edges.at(0)?.node.id,
+      preferedTactic: HeistPreferedTacticEnum.Loud,
+      difficulty: HeistDifficultyEnum.Normal,
+      minimumPayout: 100000,
+      maximumPayout: 1000000,
+      allowedEmployees: employeesFormatted.filter(
+        (employee) => employee.establishmentId === establishments.edges[0].node.id,
       ),
+      forbiddenUsers: [],
+      forbiddenAssets: [],
+      objectives: [],
     },
   });
 
+  // Watch when the establishment changes to update the employees options
+  const watchEstablishment = methods.watch('establishment') as Option | string;
+
+  // Get the current establishment (string or Option)
+  const currentEstablishment =
+    typeof watchEstablishment === 'string' ? watchEstablishment : watchEstablishment?.value;
+
+  // Get the employees of the current establishment
+  const [allowedEmployeesOptions, setAllowedEmployeesOptions] = useState<
+    (Option & { establishmentId: string })[]
+  >(employeesFormatted.filter((employee) => employee.establishmentId === currentEstablishment));
+
+  // Update the employees options when the establishment changes
+  useEffect(() => {
+    setAllowedEmployeesOptions(
+      employeesFormatted.filter((employee) => employee.establishmentId === currentEstablishment),
+    );
+
+    methods.setValue('allowedEmployees', allowedEmployeesOptions);
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [watchEstablishment]);
+
   return (
-    <div>
+    <>
       <Dialog.Title asChild>
         <Heading as="h2" size="8">
-          {t('edit')} - {heist.name}
+          {t('add')}
         </Heading>
       </Dialog.Title>
-      <Section className="space-y-3" size="1">
+      <Section size="1">
         <RemixFormProvider {...methods}>
           <form
-            id="heist-edit-form"
+            id="heist-add-form"
             method="post"
             className="space-y-4"
             onSubmit={methods.handleSubmit}
           >
-            <FieldInput name="name" label={t('name')} type="text" required />
+            <FieldInput name="name" label={t('name')} type="text" />
             <TextAreaInput name="description" label={t('description')} />
+
             <Grid columns="2" gap="2">
               <FieldInput name="startAtDate" label={t('heist.start_at.date')} type="date" />
               <FieldInput name="startAtTime" label={t('heist.start_at.time')} type="time" />
             </Grid>
+
             <Grid columns="2" gap="2">
               <FieldInput
                 name="shouldEndAtDate"
@@ -291,14 +255,22 @@ export default function Edit() {
                 type="time"
               />
             </Grid>
+
             <Grid columns="2" gap="2">
               <FieldInput name="minimumPayout" label={t('heist.minimum_payout')} type="number" />
               <FieldInput name="maximumPayout" label={t('heist.maximum_payout')} type="number" />
             </Grid>
+
+            <FieldSelect
+              name="establishment"
+              label={t('establishment')}
+              options={establishmentsFormatted}
+            />
             <FieldMultiSelect
               name="allowedEmployees"
               label={t('heist.allowed_employees')}
-              options={employeesFormatted}
+              options={allowedEmployeesOptions}
+              disabled={!watchEstablishment}
             />
             <FieldMultiSelect
               name="forbiddenUsers"
@@ -320,11 +292,6 @@ export default function Edit() {
               label={t('heist.difficulty')}
               options={heistDifficulties}
             />
-            <FieldSelect
-              name="visibility"
-              label={t('heist.visibility')}
-              options={heistVisibilities}
-            />
             <FieldInputArray
               name="objectives"
               label={t('heist.objective')}
@@ -334,9 +301,6 @@ export default function Edit() {
                   name: '',
                   description: '',
                   optional: false,
-                },
-                add: {
-                  text: t('heist.add_objective'),
                 },
                 fields: [
                   {
@@ -357,25 +321,27 @@ export default function Edit() {
                 ],
               }}
             />
-            {methods.getValues('visibility') === HeistVisibilityEnum.Public ? (
+            <Flex gap="2">
+              <Link to={`/map/${placeId}`}>
+                <div className="flex h-8 items-center rounded-2 bg-gray-9 px-3 text-[white]">
+                  <ArrowLeftIcon />
+                </div>
+              </Link>
               <FormAlertDialog
-                title={t('heist.edit.confirm')}
-                description={t('heist.edit.confirm_description')}
+                title={t('add')}
+                description={t('heist.add.confirm')}
                 actionColor="green"
                 cancelText={t('cancel')}
-                formId="heist-edit-form"
+                formId="heist-add-form"
               >
-                <Button type="button" color="green">
-                  {t('update')}
+                <Button className="grow" type="button" color="jade">
+                  {t('create')}
                 </Button>
               </FormAlertDialog>
-            ) : (
-              <SubmitButton text={t('update')} color="green" />
-            )}
+            </Flex>
           </form>
         </RemixFormProvider>
-        <Link to={`/map/${placeId}`}>{t('back')}</Link>
       </Section>
-    </div>
+    </>
   );
 }
